python-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyodbc==5.0.1
pydantic[email]==2.5.0
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0